requirements.txt

Flask
dvc
pandas
notebook
numpy
matplotlib
seaborn
python-box
pyYAML
tqdm
ensure
joblib
types-PyYAML
scipy>=1.9.1
black

-e .
